+++ RadiAIDD.egg-info/SOURCES.txt
@@ -1,0 +1,86 @@
+.gitignore
+LICENSE
+Readme.md
+matplotlibwidgetFile.py
+ressources_rc.py
+setup.cfg
+setup.py
+.github/PULL_REQUEST_TEMPLATE.md
+.github/workflows/test.yml
+RadiAIDD/__init__.py
+RadiAIDD/__main__.py
+RadiAIDD/matplotlibwidgetFile.py
+RadiAIDD/ressources_rc.py
+RadiAIDD.egg-info/PKG-INFO
+RadiAIDD.egg-info/SOURCES.txt
+RadiAIDD.egg-info/dependency_links.txt
+RadiAIDD.egg-info/entry_points.txt
+RadiAIDD.egg-info/requires.txt
+RadiAIDD.egg-info/top_level.txt
+RadiAIDD/Backend/Children.py
+RadiAIDD/Backend/Containers.py
+RadiAIDD/Backend/IsoCenter.py
+RadiAIDD/Backend/IsoCenter5.py
+RadiAIDD/Backend/Landmark.py
+RadiAIDD/Backend/Landmark5.py
+RadiAIDD/Backend/MotorControl.py
+RadiAIDD/Backend/Radiography.py
+RadiAIDD/Backend/Registration.py
+RadiAIDD/Backend/Report.py
+RadiAIDD/Backend/Target.py
+RadiAIDD/Backend/Target5.py
+RadiAIDD/Backend/__init__.py
+RadiAIDD/Backend/matplotlibwidgetFile.py
+RadiAIDD/Backend/matplotlibwidgetFileSmall.py
+RadiAIDD/Backend/OWIS/PS_10_Betriebsanleitung_2014.pdf
+RadiAIDD/Backend/OWIS/owis.ini
+RadiAIDD/Backend/OWIS/owis_com_test.py
+RadiAIDD/Backend/OWIS/ps10.dll
+RadiAIDD/Backend/OWIS/ps10.h
+RadiAIDD/Backend/OWIS/ps10.lib
+RadiAIDD/Backend/OWIS/ps10tool.c
+RadiAIDD/Backend/Style/Darkorange.qss
+RadiAIDD/Backend/Style/stylefile.qss
+RadiAIDD/Backend/Style/QTDark-master/LICENSE
+RadiAIDD/Backend/Style/QTDark-master/QTDark.qss
+RadiAIDD/Backend/Style/QTDark-master/README.md
+RadiAIDD/Backend/Style/QTDark-master/screenshots/palette.png
+RadiAIDD/Backend/Style/QTDark-master/screenshots/screenshot_1.png
+RadiAIDD/Backend/UI/IsoCenter.ui
+RadiAIDD/Backend/UI/IsoCenter5.py
+RadiAIDD/Backend/UI/Landmark.ui
+RadiAIDD/Backend/UI/Landmark5.py
+RadiAIDD/Backend/UI/Positioning_Assistant_GUI.py
+RadiAIDD/Backend/UI/Positioning_Assistant_GUI.ui
+RadiAIDD/Backend/UI/Target.ui
+RadiAIDD/Backend/UI/__init__.py
+RadiAIDD/Backend/UI/ressources.qrc
+RadiAIDD/Backend/UI/Icons/35240-200.png
+RadiAIDD/Backend/UI/Icons/Icon_3.jpg
+RadiAIDD/Backend/UI/Icons/Icon_3.png
+RadiAIDD/Backend/UI/Icons/Pic.jfif
+RadiAIDD/Backend/UI/Icons/Pic.jpg
+RadiAIDD/Backend/UI/Icons/Pic.png
+RadiAIDD/Backend/UI/Icons/Pic.tif
+RadiAIDD/Backend/UI/Icons/Thumbs.db
+RadiAIDD/Backend/UI/Icons/cursor.bmp
+RadiAIDD/Backend/UI/Icons/move_down.png
+RadiAIDD/Backend/UI/Icons/move_down_fast.png
+RadiAIDD/Backend/UI/Icons/move_left.png
+RadiAIDD/Backend/UI/Icons/move_left_fast.png
+RadiAIDD/Backend/UI/Icons/move_right.png
+RadiAIDD/Backend/UI/Icons/move_top.png
+RadiAIDD/Backend/UI/Icons/move_top_fast.png
+RadiAIDD/imgs/Accepted_registration.PNG
+RadiAIDD/imgs/Define_button.PNG
+RadiAIDD/imgs/IsoCenter.PNG
+RadiAIDD/imgs/Workflow.gif
+RadiAIDD/imgs/launch_prompt1.png
+RadiAIDD/imgs/launch_prompt2.png
+misc/RadiAIDD_docsheet.pdf
+n2v/config.json
+n2v/events.out.tfevents.1618244235.V15NITRO
+n2v/info.txt
+n2v/validation_loss.png
+n2v/weights_best.h5
+n2v/weights_last.h5
