+++ RadiAIDD.egg-info/entry_points.txt
@@ -1,5 +1,3 @@
-{
-"gui_scripts": [
-"RadiAIDD = RadiAIDD.__main__:main"
-]
-}
+[gui_scripts]
+RadiAiDD = RadiAIDD:run
+
