--- RadiAIDD/RadiAIDD.egg-info/SOURCES.txt
@@ -1,85 +1,0 @@
-.gitignore
-LICENSE
-MAIN.py
-RadiAID.bat
-Readme.md
-environment.yml
-matplotlibwidgetFile.py
-requirements.txt
-ressources_rc.py
-setup.cfg
-setup.py
-.github/PULL_REQUEST_TEMPLATE.md
-.github/workflows/test.yml
-Backend/Children.py
-Backend/Containers.py
-Backend/IsoCenter.py
-Backend/IsoCenter5.py
-Backend/Landmark.py
-Backend/Landmark5.py
-Backend/MotorControl.py
-Backend/Radiography.py
-Backend/Registration.py
-Backend/Report.py
-Backend/Target.py
-Backend/Target5.py
-Backend/__init__.py
-Backend/matplotlibwidgetFile.py
-Backend/matplotlibwidgetFileSmall.py
-Backend/OWIS/PS_10_Betriebsanleitung_2014.pdf
-Backend/OWIS/owis.ini
-Backend/OWIS/owis_com_test.py
-Backend/OWIS/ps10.dll
-Backend/OWIS/ps10.h
-Backend/OWIS/ps10.lib
-Backend/OWIS/ps10tool.c
-Backend/Style/Darkorange.qss
-Backend/Style/stylefile.qss
-Backend/Style/QTDark-master/LICENSE
-Backend/Style/QTDark-master/QTDark.qss
-Backend/Style/QTDark-master/README.md
-Backend/Style/QTDark-master/screenshots/palette.png
-Backend/Style/QTDark-master/screenshots/screenshot_1.png
-Backend/UI/IsoCenter.ui
-Backend/UI/IsoCenter5.py
-Backend/UI/Landmark.ui
-Backend/UI/Landmark5.py
-Backend/UI/Positioning_Assistant_GUI.py
-Backend/UI/Positioning_Assistant_GUI.ui
-Backend/UI/Target.ui
-Backend/UI/__init__.py
-Backend/UI/ressources.qrc
-Backend/UI/Icons/35240-200.png
-Backend/UI/Icons/Icon_3.jpg
-Backend/UI/Icons/Icon_3.png
-Backend/UI/Icons/Pic.jfif
-Backend/UI/Icons/Pic.jpg
-Backend/UI/Icons/Pic.png
-Backend/UI/Icons/Pic.tif
-Backend/UI/Icons/Thumbs.db
-Backend/UI/Icons/cursor.bmp
-Backend/UI/Icons/move_down.png
-Backend/UI/Icons/move_down_fast.png
-Backend/UI/Icons/move_left.png
-Backend/UI/Icons/move_left_fast.png
-Backend/UI/Icons/move_right.png
-Backend/UI/Icons/move_top.png
-Backend/UI/Icons/move_top_fast.png
-RadiAIDD.egg-info/PKG-INFO
-RadiAIDD.egg-info/SOURCES.txt
-RadiAIDD.egg-info/dependency_links.txt
-RadiAIDD.egg-info/requires.txt
-RadiAIDD.egg-info/top_level.txt
-imgs/Accepted_registration.PNG
-imgs/Define_button.PNG
-imgs/IsoCenter.PNG
-imgs/Workflow.gif
-imgs/launch_prompt1.png
-imgs/launch_prompt2.png
-misc/RadiAIDD_docsheet.pdf
-n2v/config.json
-n2v/events.out.tfevents.1618244235.V15NITRO
-n2v/info.txt
-n2v/validation_loss.png
-n2v/weights_best.h5
-n2v/weights_last.h5
