+++ setup.cfg
@@ -26,10 +26,6 @@ classifiers =
 packages = find:
 python_requires = >=3.7
 
-[options.entry_points]
-gui_scripts =
-    RadiAiDD = RadiAIDD.__main__:main"
-
 
 # add your package requirements here
 install_requires =
@@ -38,3 +34,7 @@ install_requires =
     pyserial
     tqdm
     matplotlib
+
+[options.entry_points]
+gui_scripts  =
+    RadiAiDD = RadiAIDD:run
